Splash Screen 작업 (임시 이미지 사용)
